requirement.txt
pandas
numpy
matplotlib.pyplot
seaborn
streamlit
base64
from sklearn.cluster import KMeans
from sklearn.cluster import AgglomerativeClustering
import scipy.cluster.hierarchy as sch
from sklearn.metrics import silhouette_score
